Implement programatic build for meetup pages and refactor vlad meetup
