+++ src/pages/community/meetups/5ce811c728bf226ebe872d74.md
@@ -1,0 +1,13 @@
+---
+city: Berlin
+instanceID: 5ce811c728bf226ebe872d74
+---
+
+This time a deep-dive into GraphQL and serverless with Hasura.io, free and open source engine lead by Vladimir Novick.
+
+We will learn not only how to create our own GraphQL API and consume it in our Frontend apps, but also how to create serverless business logic and even combine various GraphQL servers together.
+Vladimir is Developer Advocate at Hasura.io, Google Developer Expert, consultant, worldwide speaker, published author, host of the 3factorRadio podcast, and OSS contributor.
+
+Vladimir brings years of experience with the JavaScript ecosystem and has been coding through the rise of the web. Currently, Vladimir works with Web, Mobile, VR, AR, and IoT technologies and advocates the use of GraphQL and serverless architectures as well as functional languages such as ReasonML.
+
+The format will be around 2.5 hours of talk and live coding with half an hour break in between. There'll also be 1 or 2 lightning talks at the beginning.
